requirements.txt
alembic==1.10.3
certifi==2022.12.7
click==8.1.3
cloudinary==1.32.0
dnspython==2.3.0
email-validator==2.0.0.post1
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyMySQL==1.0.3
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
urllib3==1.26.15
uuid==1.30
Werkzeug==2.2.3
WTForms==3.0.1
wtforms-validators==1.0.0
